adicionando rota para histórico e botões de pedido e histórico já redirecionando para devidas páginas
