+++ src/pages/HistoricPage.js
@@ -1,0 +1,8 @@
+import React from "react";
+import { NavigationKitchen } from "../components/kitchen/NavKitchen.js"
+
+export const HistoricArea = () => {
+  return (
+    <NavigationKitchen />
+  )
+}
